+++ Sources/KeyboardKit/Callouts/Views/CalloutButtonArea.swift
@@ -1,0 +1,74 @@
+//
+//  CalloutButtonArea.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-09-30.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+public struct CalloutButtonArea: View {
+    
+    public init(
+        frame: CGRect,
+        style: CalloutStyle) {
+        self.frame = frame
+        self.style = style
+    }
+    
+    private let frame: CGRect
+    private let style: CalloutStyle
+    
+    public var body: some View {
+        HStack(alignment: .top, spacing: 0) {
+            calloutCurve.rotation3DEffect(.degrees(180), axis: (x: 0, y: 1, z: 0))
+            buttonBody
+            calloutCurve
+        }
+    }
+}
+
+private extension CalloutButtonArea {
+    
+    var backgroundColor: Color { style.backgroundColor }
+    
+    var buttonFrame: CGRect {
+        frame.insetBy(
+            dx: buttonInset.width,
+            dy: buttonInset.height)
+    }
+    
+    var buttonInset: CGSize { style.buttonOverlayInset }
+    
+    var cornerRadius: CGFloat { style.cornerRadius }
+    
+    var curveSize: CGSize { style.curveSize }
+}
+
+private extension CalloutButtonArea {
+    
+    var buttonBody: some View {
+        CustomRoundedRectangle(bottomLeft: cornerRadius, bottomRight: cornerRadius)
+            .foregroundColor(backgroundColor)
+            .frame(buttonFrame.size)
+    }
+    
+    var calloutCurve: some View {
+        CalloutCurve()
+            .frame(width: curveSize.width, height: curveSize.height)
+            .foregroundColor(backgroundColor)
+    }
+}
+
+struct CalloutButtonArea_Previews: PreviewProvider {
+    
+    static var previews: some View {
+        CalloutButtonArea(
+            frame: CGRect(x: 0, y: 0, width: 50, height: 50),
+            style: .standard)
+            .padding(30)
+            .background(Color.gray)
+            .cornerRadius(20)
+    }
+}