+++ Sources/KeyboardKit/Styles/InputCalloutStyle.swift
@@ -12,11 +12,6 @@ import SwiftUI
  This can be used to style `InputCallout` views, by applying
  the `inputCalloutStyle` view modifier.
  
- Note that the `calloutSize` width is not the total width of
- the callout, but the total size beyond the button area. The
- width is split on both sides. So the standard `26` width is
- causing the callout to go `13` points on each side.
- 
  You can modify the `.standard` style instance to change the
  standard global style of all `InputCallout` views.
  */
@@ -24,7 +19,7 @@ public struct InputCalloutStyle {
     
     public init(
         callout: CalloutStyle = .standard,
-        calloutSize: CGSize = CGSize(width: 26, height: 50),
+        calloutSize: CGSize = CGSize(width: 65, height: 60),
         font: Font = Font.largeTitle.weight(.light)) {
         self.callout = callout
         self.calloutSize = calloutSize